Proof of concept on new particles/render loop
I made a different particle system that uses particleSystems from
THREE.js (they are point sprites) and wanted to see how it would work
into our program when I apply a bit of stress testing.  500 particles
each with a unique, updating velocity, and four if checks per particle
for out of bounds stuff, all doing alpha blending.  Required breaking
the rendering into a multipass because sprites always get drawn last and
alphablending wouldnt work if sprites were behind particles, so I render
the background, and then the foreground over it ( a stencil test would
probably speed this up, but for now it works).  Using particle systems
like this (point sprites that the gpu takes care of, is the way to go I
think.